generalize cidr parsing and improve lua tests
