+++ internal/ingress/annotations/ratelimit/main_test.go
@@ -17,8 +17,6 @@ limitations under the License.
 package ratelimit
 
 import (
-	"reflect"
-	"sort"
 	"testing"
 
 	api "k8s.io/api/core/v1"
@@ -82,23 +80,6 @@ func TestWithoutAnnotations(t *testing.T) {
 	_, err := NewParser(mockBackend{}).Parse(ing)
 	if err != nil {
 		t.Error("unexpected error with ingress without annotations")
-	}
-}
-
-func TestParseCIDRs(t *testing.T) {
-	cidr, _ := parseCIDRs("invalid.com")
-	if cidr != nil {
-		t.Errorf("expected %v but got %v", nil, cidr)
-	}
-
-	expected := []string{"192.0.0.1", "192.0.1.0/24"}
-	cidr, err := parseCIDRs("192.0.0.1, 192.0.1.0/24")
-	if err != nil {
-		t.Errorf("unexpected error %v", err)
-	}
-	sort.Strings(cidr)
-	if !reflect.DeepEqual(expected, cidr) {
-		t.Errorf("expected %v but got %v", expected, cidr)
 	}
 }
